Created main function which utilises the states and code_loader code.
